feat (ch-edc): refactoring modules extension and core
