+++ clearing-house-edc/build.gradle.kts
@@ -1,0 +1,25 @@
+/*
+ *  Copyright (c) 2022 Microsoft Corporation
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Microsoft Corporation - initial implementation
+ *
+ */
+
+plugins {
+    `java-library`
+}
+
+val javaVersion: String by project
+
+java {
+   toolchain {
+      languageVersion = JavaLanguageVersion.of(javaVersion)
+   }
+}
