+++ clearing-house-edc/core/src/main/java/de/truzzt/clearinghouse/edc/app/delegate/AppSenderDelegate.java
@@ -1,0 +1,21 @@
+/*
+ *  Copyright (c) 2022 sovity GmbH
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       sovity GmbH - initial API and implementation
+ *
+ */
+package de.truzzt.clearinghouse.edc.app.delegate;
+
+import okhttp3.ResponseBody;
+
+public interface AppSenderDelegate<R, P> {
+
+    P parseResponseBody(ResponseBody responseBody);
+}
