+++ clearing-house-edc/core/src/main/java/de/truzzt/clearinghouse/edc/types/clearinghouse/TokenFormat.java
@@ -1,0 +1,35 @@
+/*
+ *  Copyright (c) 2021 Microsoft Corporation
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Microsoft Corporation - Initial implementation
+ *
+ */
+
+package de.truzzt.clearinghouse.edc.types.clearinghouse;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import org.jetbrains.annotations.NotNull;
+
+import java.net.URI;
+
+public class TokenFormat {
+
+    @JsonProperty("@id")
+    @NotNull
+    private final String id;
+
+    public TokenFormat(@NotNull URI id) {
+        this.id = id.toString();
+    }
+
+    public String getId() {
+        return id;
+    }
+}
