+++ clearing-house-edc/core/src/main/java/de/truzzt/clearinghouse/edc/types/ids/Context.java
@@ -1,0 +1,36 @@
+/*
+ *  Copyright (c) 2021 Microsoft Corporation
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Microsoft Corporation - Initial implementation
+ *
+ */
+
+package de.truzzt.clearinghouse.edc.types.ids;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import org.jetbrains.annotations.NotNull;
+
+public class Context {
+    @JsonProperty("ids")
+    @NotNull
+    private String ids;
+
+    @JsonProperty("idsc")
+    @NotNull
+    private String idsc;
+
+    public String getIds() {
+        return ids;
+    }
+
+    public String getIdsc() {
+        return idsc;
+    }
+}
