+++ clearing-house-edc/extensions/multipart/src/main/java/de/truzzt/clearinghouse/edc/multipart/MultipartController.java
@@ -1,0 +1,153 @@
+package de.truzzt.clearinghouse.edc.multipart;
+
+import de.truzzt.clearinghouse.edc.handler.Handler;
+import de.truzzt.clearinghouse.edc.dto.HandlerRequest;
+import de.truzzt.clearinghouse.edc.dto.HandlerResponse;
+import de.truzzt.clearinghouse.edc.types.TypeManagerUtil;
+import de.truzzt.clearinghouse.edc.types.ids.Message;
+
+import jakarta.ws.rs.Consumes;
+import jakarta.ws.rs.POST;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.PathParam;
+import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.core.MediaType;
+
+import jakarta.ws.rs.core.Response;
+import org.eclipse.edc.protocol.ids.spi.types.IdsId;
+import org.eclipse.edc.spi.monitor.Monitor;
+import org.glassfish.jersey.media.multipart.FormDataBodyPart;
+import org.glassfish.jersey.media.multipart.FormDataParam;
+import org.glassfish.jersey.media.multipart.FormDataMultiPart;
+import org.jetbrains.annotations.NotNull;
+
+import java.io.InputStream;
+import java.util.List;
+
+import static de.truzzt.clearinghouse.edc.util.ResponseUtil.malformedMessage;
+import static de.truzzt.clearinghouse.edc.util.ResponseUtil.messageTypeNotSupported;
+import static de.truzzt.clearinghouse.edc.util.ResponseUtil.notAuthenticated;
+import static java.lang.String.format;
+
+@Consumes({MediaType.MULTIPART_FORM_DATA})
+@Produces({MediaType.MULTIPART_FORM_DATA})
+@Path("/")
+public class MultipartController {
+
+    private static final String HEADER = "header";
+    private static final String PAYLOAD = "payload";
+    private static final String PID = "pid";
+    private static final String LOG_ID = "InfrastructureController";
+
+    private final Monitor monitor;
+    private final IdsId connectorId;
+
+    private final List<Handler> multipartHandlers;
+
+    private final TypeManagerUtil typeManagerUtil;
+
+    public MultipartController(@NotNull Monitor monitor,
+                               @NotNull IdsId connectorId,
+                               @NotNull TypeManagerUtil typeManagerUtil,
+                               @NotNull List<Handler> multipartHandlers) {
+        this.monitor = monitor;
+        this.connectorId = connectorId;
+        this.typeManagerUtil = typeManagerUtil;
+        this.multipartHandlers = multipartHandlers;
+    }
+
+    @POST
+    @Path("messages/log/{pid}")
+    public Response request(@PathParam(PID) String pid,
+                            @FormDataParam(HEADER) InputStream headerInputStream,
+                            @FormDataParam(PAYLOAD) String payload) {
+
+        // Check if header is missing
+        if (headerInputStream == null) {
+            monitor.severe(LOG_ID + ": Header is missing");
+            return Response.status(Response.Status.BAD_REQUEST)
+                    .entity(createFormDataMultiPart(malformedMessage(null, connectorId)))
+                    .build();
+        }
+
+        // Convert header to message
+        Message header;
+        try {
+            header = typeManagerUtil.parse(headerInputStream, Message.class);
+        } catch (Exception e) {
+            monitor.severe(format(LOG_ID + ": Header parsing failed: %s", e.getMessage()));
+            return Response.status(Response.Status.BAD_REQUEST)
+                    .entity(createFormDataMultiPart(malformedMessage(null, connectorId)))
+                    .build();
+        }
+
+        // Check if any required header field missing
+        if (header.getId() == null
+                || header.getType() == null
+                || header.getModelVersion() == null
+                || header.getIssued() == null
+                || header.getIssuerConnector() == null
+                || header.getSenderAgent() == null) {
+            return Response.status(Response.Status.BAD_REQUEST)
+                    .entity(createFormDataMultiPart(malformedMessage(header, connectorId)))
+                    .build();
+        }
+
+        // Check if DAT present
+        var dynamicAttributeToken = header.getSecurityToken();
+        if (dynamicAttributeToken == null || dynamicAttributeToken.getTokenValue() == null) {
+            monitor.severe(LOG_ID + ": Token is missing in header");
+            return Response.status(Response.Status.BAD_REQUEST)
+                    .entity(createFormDataMultiPart(notAuthenticated(header, connectorId)))
+                    .build();
+        }
+
+        // Check if payload is missing
+        if (payload == null) {
+            monitor.severe(LOG_ID + ": Payload is missing");
+            return Response.status(Response.Status.BAD_REQUEST)
+                    .entity(createFormDataMultiPart(malformedMessage(null, connectorId)))
+                    .build();
+        }
+
+        // Build the multipart request
+        var multipartRequest = HandlerRequest.Builder.newInstance()
+                .pid(pid)
+                .header(header)
+                .payload(payload)
+                .build();
+
+        // Send to handler processing
+        var multipartResponse = multipartHandlers.stream()
+                .filter(h -> h.canHandle(multipartRequest))
+                .findFirst()
+                .map(it -> it.handleRequest(multipartRequest))
+                .orElseGet(() -> HandlerResponse.Builder.newInstance()
+                        .header(messageTypeNotSupported(header, connectorId))
+                        .build());
+
+        // Build response
+        return Response.status(Response.Status.CREATED)
+                .entity(createFormDataMultiPart(multipartResponse.getHeader(), multipartResponse.getPayload()))
+                .build();
+    }
+
+    private FormDataMultiPart createFormDataMultiPart(Message header, Object payload) {
+        var multiPart = createFormDataMultiPart(header);
+
+        if (payload != null) {
+            multiPart.bodyPart(new FormDataBodyPart(PAYLOAD, typeManagerUtil.toJson(payload), MediaType.APPLICATION_JSON_TYPE));
+        }
+
+        return multiPart;
+    }
+
+    private FormDataMultiPart createFormDataMultiPart(Message header) {
+        var multiPart = new FormDataMultiPart();
+        if (header != null) {
+            multiPart.bodyPart(new FormDataBodyPart(HEADER, typeManagerUtil.toJson(header), MediaType.APPLICATION_JSON_TYPE));
+        }
+        return multiPart;
+    }
+
+}
