--- clearing-house-edc/extensions/multipart/src/main/java/de/truzzt/clearinghouse/edc/multipart/util/JWTUtil.java
@@ -1,12 +1,0 @@
-package de.truzzt.clearinghouse.edc.multipart.util;
-
-import java.time.LocalDateTime;
-import java.time.ZoneId;
-import java.util.Date;
-
-public class JWTUtil {
-
-    public static Date convertLocalDateTime(LocalDateTime localDateTime) {
-        return Date.from(localDateTime.atZone(ZoneId.systemDefault()).toInstant());
-    }
-}
