+++ clearing-house-edc/extensions/multipart/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -12,4 +12,4 @@
 #
 #
 
-de.truzzt.clearinghouse.edc.multipart.ClearingHouseMultipartExtension
+de.truzzt.clearinghouse.edc.multipart.MultipartExtension
