+++ clearing-house-edc/gradle.properties
@@ -1,3 +1,2 @@
 javaVersion=17
-edcGroup=org.eclipse.edc
-edcVersion=0.3.0
+auth0JWTVersion=4.2.2
